History agent to evaluator linked GenServer
