+++ lib/iex/test/iex/interaction_test.exs
@@ -111,10 +111,10 @@ defmodule IEx.InteractionTest do
 
   test "history size" do
     opts = [history_size: 3]
-    assert capture_iex("1\n2\n3\nv(1)", opts) == "1\n2\n3\n1"
+    assert capture_iex("1\n2\n3\nv(1)", opts) == "1\n2\n3\n'1\\n'"
     assert "1\n2\n3\n4\n** (RuntimeError) v(1) is out of bounds" <> _ = capture_iex("1\n2\n3\n4\nv(1)", opts)
     assert "1\n2\n3\n4\n** (RuntimeError) v(-4) is out of bounds" <> _ = capture_iex("1\n2\n3\n4\nv(-4)", opts)
-    assert "1\n2\n3\n4\n2\n** (RuntimeError) v(2) is out of bounds" <> _ = capture_iex("1\n2\n3\n4\nv(2)\nv(2)", opts)
+    assert "1\n2\n3\n4\n'2\\n'\n** (RuntimeError) v(2) is out of bounds" <> _ = capture_iex("1\n2\n3\n4\nv(2)\nv(2)", opts)
   end
 
   ## .iex file loading
